+++ cryocon_22c.py
@@ -1,4 +1,3 @@
-import ControllerBase
 from ControllerBase import LANController
 import logging
 
@@ -8,7 +7,7 @@ class cryocon_22c(LANController):
     Cryogenic controller
     """
     def __init__(self, opts):
-        self.logger = logging.getLogger(__name__)
+        self.logger = logging.getLogger(opts.name)
         self._msg_end = '\n'
         self.commands = { # these are not case sensitive
                 'identify' : '*idn?',
@@ -21,7 +20,7 @@ class cryocon_22c(LANController):
                 'setTempAUnits' : 'input a:units k',
                 'settempBUnits' : 'input b:units k',
                 }
-        super().__init__(opts)
+        super().__init__(opts, self.logger)
 
     def checkController(self):
         val = self.SendRecv(self.commands['identify'])
@@ -38,19 +37,19 @@ class cryocon_22c(LANController):
             self.logger.error('Didn\'t connect to the correct device? Mfg %s, model %s' % (mfg, model))
             return -2
         else:
-            self.SendRecv(self.commands['SetTempAUnits'])
-            self.SendRecv(self.commands['SetTempBUnits'])
+            #self.SendRecv(self.commands['SetTempAUnits'])
+            #self.SendRecv(self.commands['SetTempBUnits'])
             self.logger.info('Connected to controller successfully')
             return 0
 
     def Readout(self):
-        vals = []
+        resp = []
         stats = []
         for com in ['getTempA','getTempB','getSP1','getSP2','getLp1Pwr','getLp2Pwr']:
             val = self.SendRecv(self.commands[com])
             if val['retcode']:
                 resp.append(-1)
-                stats.append(-1)
+                stats.append(val['retcode'])
             else:
                 try:
                     if 'SP' in com:
@@ -66,5 +65,5 @@ class cryocon_22c(LANController):
                 except Exception as e:
                     self.logger.error('Could not read device! Error: %s' % e)
                     return {'retcode' : -2, 'data' : None}
-        return {'retcode' : stats, 'data' : vals}
+        return {'retcode' : stats, 'data' : resp}
